src/data_pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=2.10.5",
    "requests>=2.32.3",
    "ruff>=0.9.6",
]
